code cleanup

+++ src/bela/str_cat.cc
@@ -112,7 +112,7 @@ inline size_t FastHexToBufferZeroPad16(uint64_t val, char8_t *out) {
 }
 
 inline size_t FastHexToBufferZeroPad16(uint64_t val, wchar_t *out) {
-  for (int i = 0; i < 8; ++i) {
+  for (size_t i = 0; i < 8; ++i) {
     auto byte = (val >> (56 - 8 * i)) & 0xFF;
     auto *hex = &strings_internal::kHexTable<wchar_t>[byte * 2];
     std::memcpy(out + 2 * i, hex, 2 * sizeof(wchar_t));
